libtracker-miner,storage: do not rely on g_drive_is_media_removable()
Fixes GB#640845

We don't want to rely on the g_drive_is_media_removable() method
because it does not tell us whether a device can be disconnected from
the system but rather if a device contains a media that might be
extracted from it. In fact, this method maps the removable flag from
the kernel block device subsystem.

If we rely on g_drive_is_media_removable(), most of the USB harddrives
are considered as non removable, and are therefor won't be indexed.

This patch proposes to check whether or not the mount point is part of
the system, and if it's not, we use g_volume_can_mount() method to
check whether the filesystem can be mounted which gives us a better
clue about whether the related device is removable or not (in the way
tracker considers a device from being removable).

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@linux.intel.com>