Use compiler symbol visiblity features to hide internal functions
Previously we relied on libtool's -export-symbols-regex feature, but we
are hoping to drop GNU Autotools and GNU Libtool in soon in favour of
Meson.

Meson doesn't have an equivalent feature, instead the advice is to
control symbol visibility at compile time.

The approach taken in this patch is based on Pango's build system. Pango
tells the compiler to hide symbols by default (if possible), and then defines
a _PANGO_EXTERN variable at compile time which marks a single symbol as
public. In Pango's case there is then further machinary to hide symbols
based on deprecation policies but I have not copied that here, instead I
used _LIBMEDIAART_EXTERN directly.

If a compiler doesn't support hiding symbols then the library we build
makes all symbols available, which is exactly what would happen before
on platforms where the libtool didn't have an implementation for
-export-symbols-regex.

See also:

  http://mesonbuild.com/FAQ.html#how-do-i-do-the-equivalent-of-libtools-exportsymbol-and-exportregex
  https://git.gnome.org/browse/pango/
  https://git.gnome.org/browse/pango/tree/pango/pango-version-macros.h
  https://gcc.gnu.org/wiki/Visibility
  https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html#Code-Gen-Options (-fvisibility)

https://bugzilla.gnome.org/show_bug.cgi?id=783562