Rationalize NULL/empty string handling in media_art_strip_invalid_entities()
We return NULL if input is NULL and a newly allocated empty string if
input is "". Some comments disagreed with this, and it's possible the
change causes a memory leak in some app, but the alternative of returning
NULL when passed "" is dangerous as some code may free the return value
in this case. (In fact, libmediaart itself does so).

Also, make behaviour occur independently of whether
`G_ENABLE_CONSISTENCY_CHECKS` was defined at build time.